requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
idna==3.4
Paginator==0.5.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
idna==3.4
Paginator==0.5.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
Paginator==0.5.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
PySocks==1.7.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
wsproto==1.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
packaging==23.2
Paginator==0.5.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
packaging==23.2
Paginator==0.5.1
psycopg2-binary==2.9.9
pymemcache==4.0.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.7
django-redis==5.4.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
Paginator==0.5.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pymemcache==4.0.0
PySocks==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.23.1
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.7
webdriver-manager==4.0.1
wsproto==1.2.0
